[Design] #160 - NormalHeaderCollectionView UI 구현
